server/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dnspython==2.6.1
fastapi==0.115.0
firebase-admin==6.5.0
google-api-core==2.19.0
google-api-python-client==2.146.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.23.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httplib2==0.22.0
idna==3.10
msgpack==1.1.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.9.1
pyparsing==3.1.4
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
wheel==0.44.0
